requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
aiosqlite==0.18.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
discord==2.1.0
easy-pil==0.1.9
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
imageio-ffmpeg==0.4.7
multidict==6.0.3
numpy==1.24.0
Pillow==8.4.0
py-cord==2.3.2
pycodestyle==2.10.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
